improved client unavailable error handling
