+++ server/backend/influx/influxClient.go
@@ -2,12 +2,13 @@
 package influx
 
 import (
-	"errors"
 	"crypto/sha256"
 	"encoding/hex"
+	"errors"
 	"github.com/TUM-Dev/Campus-Backend/server/env"
 	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
 	"github.com/influxdata/influxdb-client-go/v2/api"
+	"github.com/influxdata/influxdb-client-go/v2/api/write"
 	log "github.com/sirupsen/logrus"
 	"os"
 )
@@ -29,80 +30,17 @@ func SetClient(client *influxdb2.Client) {
 	Client = client
 }
 
+func LogPoint(p *write.Point) {
+	w, err := writeAPI()
 
-func LogRegisterDevice(deviceId string) {
-	write := writeAPI()
+	if err != nil {
+		logClientNotConfigured()
+		return
+	}
 
-	hashedDeviceId := hashSha256(deviceId)
+	w.WritePoint(p)
 
-	p := influxdb2.NewPointWithMeasurement("ios_register_device").
-		AddTag("device_id", hashedDeviceId)
-
-	write.WritePoint(p)
-
-	FlushIfDevelop(write)
-}
-
-func LogRemoveDevice(deviceId string) {
-	write := writeAPI()
-
-	hashedDeviceId := hashSha256(deviceId)
-
-	p := influxdb2.NewPointWithMeasurement("ios_remove_device").
-		AddTag("device_id", hashedDeviceId)
-
-	write.WritePoint(p)
-
-	FlushIfDevelop(write)
-}
-func LogIOSNewGrades(deviceId string, gradesCount int) {
-	write := writeAPI()
-
-	hashedDeviceId := hashSha256(deviceId)
-
-	p := influxdb2.NewPointWithMeasurement("ios_new_grades").
-		AddTag("device_id", hashedDeviceId).
-		AddField("new_grades_count", gradesCount)
-}
-func LogIOSSchedulingDevicesToUpdate(devicesToUpdateCount int, priority int) {
-	write := writeAPI()
-
-	p := influxdb2.NewPointWithMeasurement("ios_scheduling_devices_to_update").
-		AddTag("priority", string(rune(priority))).
-		AddField("devices_to_update", devicesToUpdateCount)
-
-	write.WritePoint(p)
-
-	FlushIfDevelop(write)
-}
-
-func LogIOSBackgroundRequest(deviceId, requestType, reason string) {
-	write := writeAPI()
-
-	hashedDeviceId := hashSha256(deviceId)
-
-	p := influxdb2.NewPointWithMeasurement("ios_campus_token_request").
-		AddTag("device_id", hashedDeviceId).
-		AddField("request_type", requestType).
-		AddField("notification_reason_response", reason)
-
-	write.WritePoint(p)
-
-	FlushIfDevelop(write)
-}
-
-func LogIOSBackgroundRequestResponse(deviceId, requestType string) {
-	write := writeAPI()
-
-	hashedDeviceId := hashSha256(deviceId)
-
-	p := influxdb2.NewPointWithMeasurement("ios_campus_token_response").
-		AddTag("device_id", hashedDeviceId).
-		AddField("request_type", requestType)
-
-	write.WritePoint(p)
-
-	FlushIfDevelop(write)
+	flushIfDevelop(w)
 }
 
 func hashSha256(s string) string {
